revamped api functions to follow edge rules req.json() and returning new response

+++ pages/api/deleteDB.js
@@ -1,14 +1,17 @@
 import { db } from "@/db/db";
 import { transaction } from "@/db/schema";
 import { eq } from "drizzle-orm";
-export default async function handler(req, res) {
+export default async function handler(req) {
+  if (req.method !== "POST") {
+    return new Response(null, { status: 404, statusText: "Not Found" });
+  }
   try {
-    const { id } = req.body;
+    const { id } = await req.json();
     const newRes = await db.delete(transaction).where(eq(transaction.id, id));
-    res.status(200).json(newRes);
+    return new Response(JSON.stringify(newRes), { status: 200 });
   } catch (e) {
     console.log(e);
-    res.status(500).json([]);
+    return new Response(null, { status: 400, statusText: "Bad Request" });
   }
 }
 export const runtime = "edge";
